requirements.txt
absl-py==1.3.0
argcomplete==1.10.0
astunparse==1.6.3
beautifulsoup4==4.8.0
bitarray==2.6.2
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
et-xmlfile==1.1.0
extract-msg==0.23.1
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==22.11.23
fonttools==4.38.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
IMAPClient==2.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
miniful==0.0.6
murmurhash==1.0.9
mypy==0.991
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
olefile==0.46
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathy==0.10.1
pdfminer.six==20181108
pdfreader==0.1.12
Pillow==9.4.0
preshed==3.0.8
protobuf==3.19.6
pyahocorasick==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
pyFUME==0.2.25
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.18
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
simpful==2.9.0
six==1.12.0
sklearn==0.0.post1
smart-open==6.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
SpeechRecognition==3.8.1
srsly==2.4.5
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
textract==1.6.3
thinc==8.1.6
threadpoolctl==3.1.0
tika==1.24
tomli==2.0.1
torch==1.13.0
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzlocal==1.5.1
urllib3==1.26.13
wasabi==0.10.1
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.7.2
xlrd==1.2.0
XlsxWriter==3.0.5
zipp==3.10.0
